remove all=true from newtests
Summary:
this is a non-default setting, so it should not default to true for all tests. tests that want to test that setting should opt into it.

in particular, this is masking a bad test in `lsp/queries` that expects coverage to not return results for non-@flow files. will fix that next in the stack.

Reviewed By: samwgoldman

Differential Revision: D27012503 (https://github.com/facebook/flow/commit/10865c5c9d1a42d5a8668b4b98fb8994a335a52d)

fbshipit-source-id: 142db260c81b0fe19cea3e070bc56060d16d75be