[lsp] fix running `flow lsp` on symlinked roots
Summary:
if you add a workspace folder to VS Code that is a symlink, flow will realpath that `rootUri` and start/connect to a server for that instead.

this means that many of the DocumentUris we send back to the editor will be realpaths, which don't match the URIs in the workspace. many language services don't work or misbehave due to this. Go to Definition, for example, will open the realpath'd files outside the workspace, even if the same file is already open in a different tab!

we use realpaths pervasively, especially to deal with node modules, which are very often symlinked and whose dependencies are relative to the file's realpath.

for context/terminology:
- the editor is the client
- it spawns a `flow lsp` process which intermediates between the client and the server
- the `flow server` process is the server, which does most of the work. it talks to multiple clients via the `flow status` CLI or multiple `flow lsp` processes. it sometimes restarts, so the `flow lsp` process stays alive and reconnects to the new one.

this diff intercepts LSP messages and converts the DocumentUris:
1. within `flow lsp`, the LSP messages match what the client sent/expects
2. before `flow lsp` sends messages to the server, it realpath's them so the server only ever has to deal with realpaths
3. when `flow lsp` receives a message from the server, it converts any paths that are within the root back to the symlink path, so it matches what the client expects

(2) is not _so_ critical, because the server realpaths everything itself anyway (via `Path.make`).

(3) is the key fix. if `/home/foo/myproject` is a symlink to `/somewhere/else/myproject`, then `/somewhere/else/myproject/some/subpath.js` must also be reachable by `/home/foo/myproject/some/subpath.js`, so we do that prefix replacement.

--

note, however, that this only applies to the root being a symlink, not to symlinks within the project. if `..../myproject/foo` -> `..../myproject/bar`, then we will still return `bar/file.js` URIs even if you opened `foo/file.js`, and VS Code could still open both copies. we could _possibly_ keep track of all open files' realpaths and map those, but that's left for future experimentation.

further, if a symlink points out of the project root, even the root symlink replacement won't work: if `/somewhere/else/myproject/external` -> `/somewhere/else/otherproject` and `/home/foo/myproject` -> `/somewhere/else/myproject`, then the prefix doesn't match and we don't map `/somewhere/else/otherproject` back to `/home/foo/myproject/external`. even if this is possible (e.g. unique symlinks), this is just not worth the bookkeeping.

Reviewed By: vrama628

Differential Revision: D26989790 (https://github.com/facebook/flow/commit/9dd8cc62e912c08e140ee308e50e4daf73927524)

fbshipit-source-id: e200fbfa48a48f008074bb8a9e145e7b8cb085b1