[normalizer] fix order of environment's tparams list
Summary:
The order of type parameters in the normalizer's environment matters in applying
substitutions correctly before calling into Flow_js (see evaluate-type-destructors mode).

The `PolyT` case of `ExpandMembersConverter` was inconsistent with the reverse (implied) order.
In fact, the update in the environment there was wrong in either order of tparams.

This diff makes explicit the reverse order of typeparams in the normalizer environment and other
relevant places.

Reviewed By: samwgoldman

Differential Revision: D27106420

fbshipit-source-id: 3f91449346f3ddfddf3ae4341984b120caf5c254