extract flowlibs in server process instead of client
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/8618

the `flow start` command extracts libs and concurrent `flow start` calls can cause the libs to be missing or partially written when the server is starting up, leading to errors in lib files.

instead, the libs should be extracted in `flow server`.

### Background

the `flow` binary has all of its libs built into itself, and writes them to `/tmp/flow/flowlib_<hash>_<uid>` via the `Flowlib.extract` function.

currently this function is called by `CommandUtils.file_options` which is called by the `server`, `check`, `start`, and `ls` commands.

`server` and `check` make sense -- these commands actually need to read the libs and do typechecking. `start` and `ls` are client commands. they can run multiple times throughout the life of the server.

it's even expected that two `flow start` calls might race. they both call `flow server` and then the server uses `Lock` to make sure only one actually starts; the other exits with `Lock_stolen` exit code and whoever called `flow start` can go ahead and connect to the now-already-running server.

however! these two racing `flow start` calls means that if we're unlucky, the 2nd one can be in the process of re-extracting the flowlibs when the newly-started `flow server` process is trying to read them. this causes the server to have missing-lib errors that can't be resolved without `flow stop`.

### Fix

We only need the path to the libs in `flow start`. they don't need to be extracted yet. Instead, the server can extract them just before it needs to list them, in `files.ml`.

### Future work

The only reason we need the path to the libs in `flow start` is because it's part of `Options.t`, and the only reason for that is that it depends on the `--no-flowlib` flag. I think we could just store `no_flowlib` in `Options` and only care about the path in the server. I wanted to keep this minimally invasive to merge it, so let's follow up on that.

Reviewed By: nmote

Differential Revision: D27136789

fbshipit-source-id: 301bee339bb4b8cc9e9055aa31f63727a5c77b98