[libdef] move cluster$... into cluster module
Summary:
instead of declaring globals like `cluster$Worker`, it should be a `Worker` class inside the `cluster` module.

I filled in the settings objects while I was touching it, but there is still a lot missing from this module. my only goal here is to clean up the fake `cluster$...` globals.

Reviewed By: samwgoldman

Differential Revision: D27189082

fbshipit-source-id: 3b7a6f6512038cb770fc50d5affd713abc9bff91