Add check phase for new check mode
Summary:
In "normal" mode, checking a file first requires loading the signature contexts
for our (direct) dependencies, which include all type information we might
possibly need.

In new check mode, we do not use signature contexts. Instead, if we need types
from a dependency when checking a file, we materialize that type on demand, then
cache the result.

These two ways of providing dependency types to check have very different
performance qualities. Signature contexts can become huge, because long
dependency chains are copied down and flattened. This leads to excessive space
usage in the shared heap, as well as OCaml heap memory usage for the
deserialized values.

On the other hand, the types in signature contexts are computed only once, then
stored. This is a classic space/time tradeoff, where we pay a great deal of
space. However, note that the bloated signature contexts also costs time, since
serialization/deserialization isn't free, and they result in large heaps which
creates GC pressure.

Compared to signature contexts, creating types on demand means we might
construct a given type multiple times -- converted types can not be shared
between jobs, but can be shared between files in a single job. However, type
signatures have been designed to make this conversion quite efficient.

Since signature contexts contain every type we might possibly need, the end
result is that we drastically *over* fetch when deserializing them from shared
memory. Checking a file might only need a small fraction of the types contained
in a dependency's signature context. In this case, on-demand materialization of
types is a win in both space and time.

Currently this mode is functional but still requires some optimization to
achieve good performance in practice. This diff lays the essential groundwork,
but I will continue to iterate on performance in future work.

Reviewed By: panagosg7

Differential Revision: D27169687

fbshipit-source-id: dc5063231f985c6d6bb4b705e483aa0aeb6ced42