Refactor check APIs to support caching
Summary:
Instead of running "check" directly, we call "mk_check" which returns a function
that closes over some environment. In types-first 2.0, you might call this
environment a "check context" and it contains, primarily, a cache.

For now, a closure is sufficient to represent this check context. In the future,
it may become necessary to be more explicit. For example, we may want to provide
a capability to clear the cache. However, just letting the GC collect the
closure should be enough!

Note that we create a single closure for each check job, then re-use it for
every file in that job. This way of calling ensures that the caches are shared
between files in a job.

Regarding API design, my hope is that the combination of the `mk_` prefix and
the explicit `unit` argument will make it clear that the closure is significant.

Reviewed By: nmote

Differential Revision: D27168590

fbshipit-source-id: c4a4c3384493a0c7073b06f225f72b75f791d725