Add skeleton merge phase for new check mode
Summary:
In new check mode, we do not use signature contexts. The merge phase's primary
responsibility is to create signature contexts and store them in the heap.

The merge phase has a few other responsibilities:
1. Calculating and storing sig hashes
2. Writing to the leader heap
3. Updating suppressions, which are used to filter streaming errors

After this diff, under new check mode, the merge phase will only perform the
bulleted items (2) and (3) above, and will not store signature contexts or sig
hashes.

Recheck optimizations are not currently implemented. We always return "true" for
`diff`, indicating that the merge stream should proceed with merging
dependencies. Recheck optimizations will be implemented later, including
calculating sig hashes.

Note that the new merge code runs as part of the merge stream. The merge stream
still has logic to (a) revive context heap data and (b) write sig contexts when
a merge job raises.

For (a), this is not an issue for two reasons: (1) we won't actually try to
revive anything because recheck opts is not implemented, and (2) even if we did
try to revive sig contexts, it's a noop to revive something that doesn't exist.

(b) is not ideal, since we would rather never write sig contexts, but writing
one anyway will not cause problems except a small amount of wasted space. We can
clean this up once new check mode is the sole check mode and signature contexts
are gone.

Finally, note that new check mode is not fully implemented yet. If you try to
check any files under new check mode as of this diff, you will get crashes when
trying to read signature contexts. The next diff in this stack introduces the
new implementation for the check phase.

Reviewed By: nmote

Differential Revision: D27169190

fbshipit-source-id: 37d3c1596daf8b737537f5e65048303cf3618316