Change test which accidentally relied on recheck opts
Summary:
Because we don't include flowlibs in this test, the `x.toString()` call is an
error. The error is not included in the expected output because we don't check
this file, due to recheck optimizations. However, the purpose of this test is
not to assert anything about recheck optimizations; this behavior appears to be
unintentional.

By the principle that a test should (ideally) only break when the thing being
tested breaks, I have changed this call to a property access that is not also a
type error.

Without this change, the test starts failing under new check mode, which does
not currently support recheck optimizations. By making this change, I don't need
to disable this test in the follow-up diff which runs the automated tests under
new check mode.

Reviewed By: nmote

Differential Revision: D27169709

fbshipit-source-id: aa7267de0dba21cd118d945d75fd3fb8c6efac9b