Don't record the stats for a recheck that is part of init
Summary:
The recheck that we (under some specific circumstances) perform as part of an init is not likely to be a typical recheck, and we probably shouldn't record it as part of the recheck stats which we use to decide whether to restart or recheck.

Also, this enables the next two changes.

In practice this won't often make a difference, because this recheck runs only under rare circumstances. D27176179 makes it much clearer when this actually runs.

Reviewed By: samwgoldman

Differential Revision: D27205122

fbshipit-source-id: 33c155e1e004aba0a53d920ccce1733329f0313a