Include a recheck which is part of init in the init stats
Summary:
This recheck occurs as part of init, and as such we should account for it as part of the init time. We use these stats to determine whether to restart or recheck later on, and omitting the cost of this recheck could lead us to have a rosy picture of startup performance.

In practice this won't often make a difference, because this recheck runs only under rare circumstances. The next change makes it much clearer when this actually runs.

Reviewed By: samwgoldman

Differential Revision: D27205121

fbshipit-source-id: 5ab20568ff5c7bac80055bccea2192b92bd7a1c2