Add new directory for typechecking-related types
Summary:
I'd like to use this type in `Types_js` itself, but I can't put it there while still using it in the persistent connection, because the `inference`  directory depends on the `persistent_connection` directory, so doing so would cause a circular dependency.

The solution is to create a new directory that both can depend on.

Reviewed By: mroch

Differential Revision: D27295960

fbshipit-source-id: a1f84d166031ea9a00f434e3dba92af75f401a99