properly error when super is used outside a method
Summary: Flow does not properly error when `super` is used outside of a class method. As can be seen here: https://jsfiddle.net/uv02qr8b/7/ the behavior of `super` in arrow and function properties is not what Flow thinks it is. We should just ban referencing it in these contexts entirely.

Reviewed By: jbrown215

Differential Revision: D27291810

fbshipit-source-id: 31aff42c4a862842ae74ddf99dde37007d794c8d