properly error when super is used outside a method
Summary: Flow does not properly error when `super` is used outside of a class method. As can be seen here: https://jsfiddle.net/uv02qr8b/7/ the behavior of `super` in arrow and function properties is not what Flow thinks it is. We should just ban referencing it in these contexts entirely.

Reviewed By: jbrown215

Differential Revision: D27291810 (https://github.com/facebook/flow/commit/d563cfb4385b0561650ead8d5f421616a0c17561)

fbshipit-source-id: 3240ed6e9408c7ce24eefd66c48988afad968a22