Remove config option to disable new signatures
Summary:
I left the field in the Options record, but it is always `true` after this diff.
The old signatures code is effectively dead at this point.

While making this change, I also found and fixed a bug in the new_check parser,
which always enabled new_check mode, regardless of the actual configured value.
Oops!

Reviewed By: panagosg7

Differential Revision: D27305973

fbshipit-source-id: 1d01c3ca871f6250a37b6ba6884905ed0e091c0f