Use new signatures for annotate-exports codemod
Summary:
Before this diff, the annotate-exports codemod still used the "old" signature
builder for types-first 2.0. This diff uses the new signature builder instead.

The test output changes have two causes:
1. Objects with multiple unexpected keys
2. Missing annotations for pattern variable declarations

For (1), we ask for an annotation in the same location (the object literal) but
only emit a single verification error for the first unsupported key. I added a
comment with a TODO to potentially improve support here.

For (2), the old signatures erroneously asked for annotations on each
destructured binding, but the new signatures correctly ask for annotations on
the pattern itself. Unfortunately, we are unable to automatically add
annotations in either case, because the typed AST does not support pattern
annotations anywhere.

I think (2) can be improved by fixing the typed AST to support annotations at
the top level of patterns.

Reviewed By: panagosg7

Differential Revision: D27303552

fbshipit-source-id: 28123d542c83b9ff490f466d123f415c14502621