Further simplify parsing service code
Summary:
This diff pulls some definitions closer to the left margin. It also removes a
dependency on Signature_builder_deps, which will be removed shortly.

We used Signature_builder_deps to build up a set of signature errors. We can
remove the dependency for a simple reason: the set is unnecessary. For one, we
will not raise the same signature error twice (this may not have been true for
the old signatures, but is certainly true for the new ones). Furthermore, these
tolerable errors will be put into a set down the road. Even if we did raise the
same error twice, the only effect would be sending more data back to the server
process.

Reviewed By: panagosg7

Differential Revision: D27308649

fbshipit-source-id: 9dfdef6d159807614f2a543c23432d373c5c9732