Add test for docblock error in check-contents
Summary:
We should issue:

```
Unexpected `flow` declaration. Only one per file is allowed. [duplicate-flow-decl]
```

in both cases, but we only do so when there is a syntax error in addition to the docblock error. This is inconsistent with what `flow status` reports -- it reports the docblock error regardless.

Upcoming diffs will address this bug.

Reviewed By: mroch

Differential Revision: D27415042

fbshipit-source-id: c258655a20f344d15b931dbad0fa742b53bf10e6