stabilize order of Fuzzy_path results
Summary:
`std::priority_queue`'s comparator must provide a *strict weak ordering* (http://www.cplusplus.com/reference/queue/priority_queue/). that means `<` not `<=`.

the comparator used in `Fuzzy_path` implemented `<=`; it can tie if the scores are equal and the lengths are equal. for example, searching `["foo"; "far"; "Foo"]` for `"f"`, all 3 have the same scores and lengths.

GCC and clang implement this undefined behavior differently, which leads to different sorting on Mac and Linux.

you could argue that if the scores and lengths are the same, then it doesn't matter how it's sorted -- and that's mostly true! but that makes testing harder.

to break the tie, I sorted them lexicographically.

--

note: another option might be to maintain insertion order. I believe this could be easily done using the `lookup_` table we already keep, which maps items to their insertion-order index. it would be cheaper than `string::compare`. however, for autoimports we process files in parallel and add the results in arbitrary order, so that wouldn't work for our use case.

Reviewed By: nmote

Differential Revision: D27422330

fbshipit-source-id: 3c4997e9e2decfa1d6daf84aede6bdb83df3d090