fix race in incremental_non_flow_move test
Summary:
`incremental_non_flow_move` has been flaky on our internal Mac CI recently, failing because after the `mv tmp/foo.js .`, the `foo` module is still unknown. this suggests that Flow hasn't seen the file move yet, which is indeed a possible race.

this adds a 1s sleep to give it time.

a more ideal fix would be a command that waits for dfind to see a change, like `flow file-watcher-wait --timeout 1000 foo.js`, but does not trigger a recheck like `force-recheck` does. that is left for future work.

Reviewed By: panagosg7

Differential Revision: D27399309 (https://github.com/facebook/flow/commit/c1945d8e6e026847826a6bb9f98b9dbc89b8e512)

fbshipit-source-id: 5af4701e0a5021ad57c60e2d560fef5c4c464a15