[tool] add ability to assert exact LSP message contents
Summary:
LSP tests have this wacky behavior where it tests whether a response is expected by chunking it into tiny key:value pairs and making sure those pieces exist anywhere in the response. this makes it impossible to assert that things appear in a certain order, or that there are no additional things that we're not testing for. this is important to me for autocomplete results.

this diff makes `verifyAllLSPMessagesInStep` take an object in addition to a string. if you pass an object, it is `JSON.stringify`'d against the object from the response. since both of them are `JSON.stringify`'d, whitespace doesn't matter.

Reviewed By: nmote, vrama628

Differential Revision: D27331866 (https://github.com/facebook/flow/commit/0ebc38a775d12cb3ede2c0e70a9c59f8871d1270)

fbshipit-source-id: 68a19ff4f105598e0ddd17d683369856511a52d5