[lsp] fix missing close() after shutdown()
Summary:
before we used ocamlformat, this was written as

```
try Timeout.shutdown_connection env.c_conn.ic with _ -> ();
try Timeout.close_in_noerr env.c_conn.ic with _ -> ()
```

the intention was to try to `shutdown()` and then `close()`, but this is actually parsed as

```
try Timeout.shutdown_connection env.c_conn.ic
  with _ ->
    ();
    (try Timeout.close_in_noerr env.c_conn.ic with _ -> ())
```

I don't know whether it's necessary to shutdown if we're about to close anyway, but the reverse is true: `shutdown_connection` only closes the send side, so we still need to `close()` after. thus we were likely leaking FDs.

https://beej.us/guide/bgnet/html/#close-and-shutdownget-outta-my-face

Reviewed By: samwgoldman

Differential Revision: D27431833

fbshipit-source-id: 80eda093f694e67ec08e50060381c201bc69b752