[tool] better fix for doesMessageMatch
Summary:
D27331866 (https://github.com/facebook/flow/commit/0ebc38a775d12cb3ede2c0e70a9c59f8871d1270) added the ability to test LSP messages using an object literal instead of substring matching, but it compared the entire actual `LSPMessage` with just the expected result, e.g. `{method: "textDocument/definition", result: {foo: "bar"}}` vs `{foo: "bar"}`.

this always failed, but when it added to the `actuals` array that's used in the diff, it only added the `result` value, so the test ends up working correctly for the wrong reason.

additionally, there is no way to use these new object literals for notifications (which have a `params` prop instead of a `result` prop) or for errors (which have an `error` prop instead).

to fix, pass in an expected `LSPMessage` instead of the method name and result separately. see the few updated tests as examples.

--

as a bonus, this almost fixes the ability to use `tool record` on these tests! the only remaining problem with `tool record` is that it inlines `['textDocument/publishDiagnostics', ...lspIgnoreStatusAndCancellation]` into `['textDocument/publishDiagnostics', 'window/showStatus', '$/cancelRequest']`

Reviewed By: nmote

Differential Revision: D27372809 (https://github.com/facebook/flow/commit/5f71867e708ad0684695b1830f44d959401b0bb8)

fbshipit-source-id: 4daf350c3daf9dec7c55d0a3dd0690e0984e3227