Add test for module-type-conflict error in check-contents
Summary:
This error is generated by `File_sig`, and shows up as a strange special case in `Parsing_service_js.do_parse`. If the `fail` parameter is set to `false`, it's the only way that `do_parse` can return `Parse_fail` except if it encounters an unexpected exception (which is caught and converted into a parse error).

I will shortly be making some changes to much of this logic, so I want to make sure that all the edge cases are covered by tests.

Reviewed By: samwgoldman

Differential Revision: D27416809

fbshipit-source-id: 7476f0c7fb51b79096f4dccdaa49780d8ff4693a