Lazy assert export is type
Summary:
Before this change, forcing the exports of a dependency in new check mode would
eagerly force any exported types as well, due to the AssertExportIsTypeT
constraint.

This diff makes that constraint lazy, so it will only be emitted when the
exported type is actually used in a dependent module.

Reviewed By: panagosg7

Differential Revision: D27383864

fbshipit-source-id: 661befbd0715f5caba55be97fccd97bfac4e0641