Move AssertExportIsTypeT to statement.ml
Summary:
Before this diff, the act of exporting a type would emit a constraint on the
exported type. This constraint errors if the exported type is not actually a
type and converts the export to `any`.

This diff moves that assertion constraint into statement.ml instead. I also
updated Type_sig_merge to emit the same constraint when calculating the exports
from a signature.

Why do this? Well, consider code like this:

```
// a.js
type T = string
type U = ... massive type annotation ...
export type {T, U}

// b.js
import type {T} from './a';
```

When checking module B in new check mode, the types for module A will be
constructed lazily and on demand. Both T and U are lazy thunks evaluating to
types. We also want the exports of A to be a type containing those lazy thunks.

We don't want to force T and U until they are used in B, but the
AssertExportIsTypeT constraint has the effect of forcing both T and U. In a
follow-up, I will address this by making the assertion "lazy" in the signature
as well, a la D27382487 and D27382488.

Reviewed By: panagosg7

Differential Revision: D27447555

fbshipit-source-id: 9f2d18e3b7c9d35461fd201ce630bbd4bcf9765b