Lazy repositioning
Summary:
We perform repositioning in Type_sig_merge. This can lead to poor performance in
new check mode by unnecessarily forcing lazy types.

This diff makes the repositioning lazy. That is, instead of performing the
repositioning when we convert from a signature to a type, we do it when that
type is used downstream.

Reviewed By: panagosg7

Differential Revision: D27382487

fbshipit-source-id: 05a70e02726fd5b7b439e934662bc4edc8d4eb11