Clean up export declarations
Summary:
Before this diff, the `export_statement` function (renamed `export_specifiers`
here), used to take an "export info" argument, a list of 4-tuples describing
certain export declarations.

In my opinion, it was unnecessarily difficult to understand how this code
worked. The `export_from_local` function that consumes the 4-tuple is full of
conditional logic, suggesting that the inputs are not particularly regular.

This diff moves the responsibility for these export declarations closer to the
AST visitors. Despite the (presumed) intentions of the existing code to be DRY,
this is actually less code overall.

A side effect of this work is that we no longer register the internal name
`*default*` in the environment. This used to happen for export statements like
`export default function() { ... }` and `export default class { ... }`, which
used to call into `statement`, but then need to extract a type from the
environment.

Since we no longer call into `statement` with anonymous functions and classes, I
have turned these cases into asserts.

Reviewed By: panagosg7

Differential Revision: D27440892

fbshipit-source-id: 4822e8f21da51475bf4812d72b5d08741db50a35