Lazy constraints
Summary:
Before this diff, we added laziness to the type graph by making FullyResolved
constraints lazy in their type. This diff adds additional laziness. Now, the
constraints associated with each root is also lazy.

This makes it possible to construct lazy unresolved tvars as well, which we will
use in a follow-up diff to write a lazy `mk_instance` for new check mode.

Reviewed By: panagosg7

Differential Revision: D27382486

fbshipit-source-id: 9567a6f3f6101797749c62739fdb8ba06aca726a