Handle check_syntax in parse_contents rather than passing it down to do_parse
Summary:
I'm not a fan of the wildly different behavior of `Parsing_service_js.do_parse` based on a single flag. I think we should eliminate it, though I'm not sure I want to drive that all the way through at the moment. Ideally, `do_parse` would return the full set of parsing artifacts, including errors, and the callers would decide what to do with that information. As-is, it does a full parse anyway, but raises an exception if there are parse errors after the fact, if `fail` is set to `true`.

I'm not sure that I want to take on the full project of simplifying the `do_parse` API at the moment, but this is an incremental step towards that future regardless. Now, in `parse_contents` we always pass `~fail:false` and handle parse errors if there are any. Future diffs push this farther up the stack, removing the `check_syntax` flag as well.

Reviewed By: mroch

Differential Revision: D27475162

fbshipit-source-id: 1b349b852b8c00beb81b6fa120116f778e489891