Change typecheck_contents result when the file is skipped
Summary:
This happens when the input file is not a source file, for example when it is a JSON file. This changes the diff to use the Error case for this, so that we can, in later diffs, guarantee by construction that the Ok case includes type contents artifacts.

This also brings it more into alignment with `type_contents`

Reviewed By: mroch

Differential Revision: D27577172

fbshipit-source-id: 83081257e2782c1de52555ab944e79280194f3de