Cache type_parse_artifacts rather than type_contents
Summary:
This is in preparation for D27660824, which will remove `type_contents` in favor of inlining its body (similar to what D27634446 (https://github.com/facebook/flow/commit/1191da4e969b0cd1eecead55dbe10331c21a503c) did for `typecheck_contents`), and for later diffs in which we will begin caching the results `type_parse_artifacts` calls that were formerly part of `typecheck_contents`.

This means that the call into the parser will no longer be cached, but typechecking is the expensive part, so this shouldn't have a noticeable effect on performance. We can always add a separate cache for parse artifacts later if needed, anyway.

Reviewed By: samwgoldman

Differential Revision: D27659125

fbshipit-source-id: 702df46b0f07984542d97c5c7d9490ff75f84a1e