Fix in conflict of non-flow extensions with flow files
Summary:
Before, there were two issues with module resolution:

1/ when reading the "main" file of `package.json`:
```
{
    "name": "m",
    "main": "src/index.ts"
  }
```
We would resolve module `m` to `node_modules/m/src/index.ts`, even if a `m.js`
file was present.

2/ An import from `'./d3.svg'` would fail even though file `./d3.svg.js` is present.

This diff changes `path_if_exists` to also perform the "is-flow" check. This has
two implications:

1/ Requires the "main" property of package.json to be a valid flow extension for the file
to be considered.

2/ The update in `resolve_relative` fixes the case where a relative reference `./a.svg`
did not resolve to an existing file `./a.svg.js`, because it was considered a resource
file without fall-back to the case of a '.js' ending.

Reviewed By: mroch

Differential Revision: D27634059

fbshipit-source-id: 4bb924fee6ea956322648d779271565de758aa24