[new-env] Provider analysis and testing
Summary:
This diff determines which expressions count as providers for variables.

This is a two-pass process: first, we look at variable *declarations* and see if they're initialized -- this is necessary because, even if a declaration comes after a write, the initialization on the declaration should take precedence for figuring out the variable's type.

Second, we find all the writes to variables, and if they're not already fully provided by their initializers, we add the first write(s) as providers.

This produces a fairly complicated environment data structure, which partially exists because of the need to track information at different depths on both passes of the AST. The environment is a stack of scopes, but each scope knows about its own child scopes, stored as a loc-keyed map inside the scope. This allows us to re-enter a a scope and allows scopes to persist, rather than being a transient data structure as in the typechecking environment.

Reviewed By: panagosg7

Differential Revision: D27036150

fbshipit-source-id: b84d148690d8483f00d5c0a68a71feb84e989794