microoptimize resolving deps
Summary: use `Set.union` instead of `Set.add` to combine parsed and unparsed filenames. before, we were looping over an unsorted array of unsorted filenames and adding to two sets. now, we're adding to only one set and using it with `Set.union`. set union is more efficient because the two sets are already ordered so there is less rebalancing.

Reviewed By: gkz

Differential Revision: D27711359

fbshipit-source-id: d4009ec652f3cd2351c704739d132ca31d74c844