[enums] `getName` method runtime implementation
Summary:
The runtime implementation of the `getName` method. Given a value of the enum type, this returns the string representation of the name of the member.

There have been a variety of user requests for this feature (see section in https://fb.quip.com/JgubAOrxqMtj), which come down to:
- It's useful for internal tools to create CRUD UIs, e.g. like a dropdown list or tabs. (for external products we need to use Fbt for the labels)
- It's useful for debugging

After this lands and I update usages of the runtime, I will update the type of

https://www.internalfb.com/intern/diffusion/FBS/browsefile/master/fbcode/flow/lib/core.js?commit=076e14a43544b04a372f71d19f9e956750513f56&lines=2447%2C2448%2C2450%2C2449%2C2451%2C2452

to include

```
  getName(value: TEnum): string,
```

Reviewed By: pieterv

Differential Revision: D27728469

fbshipit-source-id: 760df8ae41fd6d749fa81d58818ba4f85e42542d