Rename make_parse_artifacts_and_errors to parse_contents
Summary: I think this naming makes it a bit clearer to clients what is going on. The only reason I didn't use `parse_contents` to begin with was that it was already in use in this file, but I renamed the old `parse_contents` to address this issue. I think the new name is more descriptive and I added a comment as well.

Reviewed By: samwgoldman

Differential Revision: D27766642

fbshipit-source-id: ba8e7a963ce21555bbfe9b57b30f256cd8cd3f25