[CircleCI] use binutils 2.35 on Windows
Summary:
binutils 2.36 pokes a limitation in ocaml for windows that breaks dynlinking, which we use for our ppx plugins.

cygwin added 2.36 on 4/10, which is why this suddenly broke in 0.149 when the version bump busted our caches. this diff hackily downgrades back to binutils 2.35.

cygwin's CLI doesn't provide a way to install old packages, so I manually download the binaries and extract them on top of the cygwin directory. this would be a bad idea to do to a normal cygwin install, but since we're doing it in a CircleCI image, it doesn't matter that cygwin doesn't know the state of its packages.

ref https://github.com/facebook/flow/issues/8645

Pull Request resolved: https://github.com/facebook/flow/pull/8644

Reviewed By: samwgoldman

Differential Revision: D27843998 (https://github.com/facebook/flow/commit/740fc5dada413af0d0b67167dd3c22754421e9e9)

Pulled By: mroch

fbshipit-source-id: db3d3172a42305ace88f83dfa5f876ea23af5753