Convert refinement scope to a stack
Summary:
We'll need to negate new refinements when handling logical OR expressions.

In order to keep new refinements separate from existing ones, we introduce new refinement scopes onto our stack of refinement scopes. This allows us to negate only the refinements that are present in a new refinement scope.

No methods ending in `_refinement` should introduce new scopes that are not popped before the method ends. Instead, they should do whatever they need to do to the new refinements and then merge that scope into the existing scope.

See D27608557 for a more thorough description.

When we read a variable, we AND together all entries into the refinement scopes.

Reviewed By: panagosg7

Differential Revision: D27608544

fbshipit-source-id: c6ebda2c83f1bee6bde5c2799554d6cd0b94712d