Scope logical refinements
Summary:
We have nothing that actually introduces refinements yet. This diff implements the refinement scoping on logical expressions,  which will allow us to start actually recording refinements.

The rest of the diffs in this stack will introduce various `expressionkind_refinement` methods, which will mostly mirror the behavior of the regular `expressionkind` methods. For example, a previous diff already added `expression_refinement`, where an existing `expression` method is provided by the flow ast mapper.
The `_refinement` versions of these methods will also add new refinements in scope that are present when the method returns. non-`refinement` versions of these methods might call the `refinement` version, but they MUST restore the original refinements before returning.  Here's how each would be called:

```
(x && y) && z;
```

When we visit this program, we have a logical expression with LHS `(x && y)` and RHS `z`. When we check `z`, we want any refinements introduced from the LHS in scope, so we call `expression_refinement` on the LHS and then  `expression` on the RHS, which will allow us to use refinements introduced from the LHS on the RHS. After checking z, we restore the original refinements so that none are in scope when we proceed on to the next statement.

When we check the LHS from expression_refinement, we'll call `logical_refinement` to check `x && y` (this is implemented in a later diff). This will do exactly what `logical` does, except it will keep the refinements introduced by x && y in scope so that when we check z we have those refinements!

Right now, the only way to introduce a refinement scope is through a logical expression, but in the future other syntax forms, like the test is an if-statement, will introduce a refinement scope. A refinement scope indicates the lifetime a refinement, and they are distinct from regular variable scopes. For example, in `(x && y) || z`, the refinement on `x` is live in `y`, but it is not live in `z` (technically, !x || !y is alive in z). The refinement on x is definitely not present in the next statement, since an expression statement does not introduce a refinement scope.

In a future diff, we will change the representation of this refinement scope into an _explicit_ stack instead of an implicit one that uses the call stack. We do this so that we can easily _negate_ refinements in refinement scopes that require it. For example, let's take `x || y`.
We push a new refinement scope and check `x`. When we check `y`, we want `Not x` in scope, so we'll negate the refinements present on the top of the refinement scope stack. If we didn't keep the explicit stack, we'd need to diff the old stack and the new to find these new refinements. An important invariant of the implementation of `_refinement` expressions is that they do not introduce new refinement scopes to the stack upon returning. They may use new refinement scopes intermediately, like we'd do in `logical_refinement`, but the refinements from the entire logical expression should be merged into the already-existing refinement scope (which must exist if we're inside a _refinement methods, since only  methods that introduce refinement scopes can call into _refinement methods). See D27608552 for more clarity on that.

Reviewed By: panagosg7

Differential Revision: D27608557

fbshipit-source-id: 476118acd17d8a02e1ec7adcd24392a1a4b5a9dd