If Vals are equal avoid merging and perturbing the Val id
Summary:
When merging ssa environments we end up calling `merge`  on every pair of Vals from the current ssa env and the given one.

In order to avoid making new IDs for entries that did not change, we do a check on the ids of the two Vals to determine whether we should even bother merging or not. Vals are always identical if their ids are identical.

Reviewed By: panagosg7

Differential Revision: D27477217

fbshipit-source-id: 71153e0d24feefc288107fdb6def0c1f7be34517