Add truthy refinement implementation
Summary:
This diff implements truthy refinements on variable names.

The strategy here is mimicked in all future diffs implementing refinements: we introduce a new `expressiontype_refinement` method that implements the normal behavior the SSA builder would have to perform on the expression and then augments it with calls to `add_refinement` when necessary.

Reviewed By: panagosg7

Differential Revision: D27608556

fbshipit-source-id: e75e4c84768a89af32607588da02a8856ebef2d9