[indexed-access] Optional Indexed Access type-checking
Summary:
Add type-checking for Optional Indexed Access (`T?.[K]`).

This is a different approach from the initial approach (D26902596) - it is much simpler and does not have any issues with implementation, and doesn't even use `OptionalChainT`. However, the type-checking behaviour differs from optional chaining at the value level in a small way: when the object is non-nullable, optional indexed access returns a type including `void`. E.g

Optional chaining:
```
const o = {a: 1};
(o?.a: number); // OK, not void

type O = {a: number};
type T = O?.['a']; // Is like `number | void` rather than just `number`
(1: ); // OK
(undefined: O?.['a']); // OK
```

One of the guiding design principles behind type destructors is that they should mirror the behaviour of their value-level equivalent, which is why I have spent a considerable amount of time trying to do this. But it does not seem possible to do given how our systems work.

In this case, I believe this small divergence in behaviour is OK, because the behaviour which we don't support is "useless". If the object type is not nullable, then using optional indexed access is not required, you can replace it with just regular indexed access for the same result.

The only reason someone would want to use optional chaining on a non-nullable object value is if they believe that Flow is lying to them and the object is actually nullable. But in this case, we still lie to them with the result type as we don't include `void`. The only safe thing for the user to do is ensure the object is typed as being nullable. Additionally, for type annotations there is no runtime behaviour that a user might want invoke, so it really doesn't make sense for them to not just use regular access.

Thus, the only reason the divergent behaviour would be a problem is if it is encountered. But if it is encountered, the user should either use non-optional chaining/access instead, or fix their object type so that is nullable.

If in the future we find a way to implement this in a way that 100% mirrors how optional chaining works at the value level, the migration will be simple. The difference will be in that in the divergent case, we will now resolve to just `T` instead of `T | void`. In order to address errors that result from this, we just need to add `| void` at the different sites that this occurs.

Reviewed By: mvitousek

Differential Revision: D27641135

fbshipit-source-id: 3ed6f6539fe5a45cd5b9aeb79177b0973f4d0fec